Dependencies:  - Recompiled Lua again...
Core:
 - Added atomics code for Windows.
 - Removed boost::atomic package.
 - Refactored log system to support multiple sinks.
 - Removed boost::noncopyable from most of the code.
 - Changed stream methods to minimize copying of data.

Editor:
 - Added log window.
 - Changed gizmos to use picking planes (WIP).
 - Added support for terrain raising of multiple cells (WIP).

+++ src/Engine/animation/Animation.cpp
@@ -99,7 +99,7 @@ Matrix4x3 Animation::getKeyFrameMatrix(const BonePtr& bone, double time)
 		const KeyFrame& keyRight = boneKeyFrames[endIndex];
 
 		float timeDelta = keyRight.time - keyLeft.time;
-		float interpolator = (time - keyLeft.time) / timeDelta;
+		float interpolator = ((float) time - keyLeft.time) / timeDelta;
 
 		position = keyLeft.position.lerp(keyRight.position, interpolator);
